+++ db/updates/update_6.2-6.3.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `ingSuppliers` ADD `address` VARCHAR(255) NULL AFTER `name`, ADD `po` VARCHAR(255) NULL AFTER `address`, ADD `country` VARCHAR(255) NULL AFTER `po`, ADD `telephone` VARCHAR(255) NULL AFTER `country`, ADD `url` VARCHAR(255) NULL AFTER `telephone`, ADD `email` VARCHAR(255) NULL AFTER `url`;