+++ pages/update_data.php
@@ -828,6 +828,12 @@ if($_GET['settings'] == 'sup'){
 if($_POST['supp'] == 'add'){
 	$description = mysqli_real_escape_string($conn, $_POST['description']);
 	$name = mysqli_real_escape_string($conn, $_POST['name']);
+	$address = mysqli_real_escape_string($conn, $_POST['address']);
+	$po = mysqli_real_escape_string($conn, $_POST['po']);
+	$country = mysqli_real_escape_string($conn, $_POST['country']);
+	$telephone = mysqli_real_escape_string($conn, $_POST['telephone']);
+	$url = mysqli_real_escape_string($conn, $_POST['url']);
+	$email = mysqli_real_escape_string($conn, $_POST['email']);
 	$platform = mysqli_real_escape_string($conn, $_POST['platform']);
 	$price_tag_start = htmlentities($_POST['price_tag_start']);
 	$price_tag_end = htmlentities($_POST['price_tag_end']);
@@ -843,18 +849,22 @@ if($_POST['supp'] == 'add'){
 	}		 
 	
 	if(empty($name)){
-		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error: </strong> Supplier name required</div>';
+		$response["error"] = '<strong>Error: </strong> Supplier name required';
+		echo json_encode($response);
 		return;
 	}
 	if(mysqli_num_rows(mysqli_query($conn, "SELECT name FROM ingSuppliers WHERE name = '$name'"))){
-		echo '<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>'.$name.'</strong> Supplier already exists!</div>';
+		$response["error"] = '<strong>'.$name.'</strong> Supplier already exists!';
+		echo json_encode($response);
 		return;
 	}
 
-	if(mysqli_query($conn, "INSERT INTO ingSuppliers (name,platform,price_tag_start,price_tag_end,add_costs,notes,min_ml,min_gr) VALUES ('$name','$platform','$price_tag_start','$price_tag_end','$add_costs','$description','$min_ml','$min_gr')")){
-		echo '<div class="alert alert-success alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a>Supplier '.$name.' added!</div>';
+	if(mysqli_query($conn, "INSERT INTO ingSuppliers (name,address,po,country,telephone,url,email,platform,price_tag_start,price_tag_end,add_costs,notes,min_ml,min_gr) VALUES ('$name','address','po','country','telephone','url','email','$platform','$price_tag_start','$price_tag_end','$add_costs','$description','$min_ml','$min_gr')")){
+		$response["success"] = 'Supplier '.$name.' added!';
+		echo json_encode($response);
 	}else{
-		echo mysqli_error($conn);
+		$response["error"] = 'Something went wrong: '.mysqli_error($conn);
+		echo json_encode($response);
 	}
 	return;
 }