+++ releasenotes.md
@@ -1,3 +1,5 @@
 Whats New in v6.3
 ----------------------
 - Add phpMyAdmin by default for docker installations, link can be found under the maintanance tab in settings
+- Complete rewrite suppliers page
+- Add additional info fields for suppliers